SourceCodeOnly/requirements.txt
requests
python-dotenv
nltk
numpy
sklearn
scipy
flask
yellowbrick